move data reductions by day for basics into basics data munger
