+++ plugins/blip/basics/chartbasicsfactory.js
@@ -68,6 +68,7 @@ var BasicsChart = React.createClass({
     var basicsData = this.props.patientData.basicsData;
     if (basicsData.sections == null) {
       basicsData = _.assign(basicsData, basicsState);
+      dataMunger.reduceByDay(basicsData);
       basicsData.data.deviceEvent.infusionSiteHistory = dataMunger.infusionSiteHistory(basicsData);
       basicsData.data.bgDistribution = dataMunger.bgDistribution(
         basicsData,
